Revert "ocr支持在抽帧时同步执行"
