+++ requirements.txt
@@ -32,5 +32,3 @@ tifffile==2021.11.2
 urllib3==1.26.7
 pysrt==1.1.2
 wordsegment==1.3.1
-tqdm==4.63.1
-multiprocess==0.70.12.2
